+++ examples/custom-types/expected/Operations/MyEnumInputQuery.php
@@ -12,7 +12,9 @@ class MyEnumInputQuery extends \Spawnia\Sailor\Operation
     /**
      * @param \Spawnia\Sailor\CustomTypes\Types\EnumInput|null $input
      */
-    public static function execute($input = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): MyEnumInputQuery\MyEnumInputQueryResult
+    public static function execute(
+        $input = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): MyEnumInputQuery\MyEnumInputQueryResult
     {
         return self::executeOperation(
             $input,