+++ examples/custom-types/expected/Operations/MyEnumInputQuery/WithEnumInput/EnumObject.php
@@ -17,8 +17,9 @@ class EnumObject extends \Spawnia\Sailor\ObjectLike
      */
     public static function make(
         $custom = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
-        $default = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'
-    ): self {
+        $default = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
+    {
         $instance = new self;
 
         $instance->__typename = 'EnumObject';