+++ examples/input/expected/Operations/TakeList.php
@@ -12,7 +12,9 @@ class TakeList extends \Spawnia\Sailor\Operation
     /**
      * @param array<int|null>|null $values
      */
-    public static function execute($values = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): TakeList\TakeListResult
+    public static function execute(
+        $values = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): TakeList\TakeListResult
     {
         return self::executeOperation(
             $values,