+++ examples/php-keywords/expected/Operations/_catch/_catch.php
@@ -13,7 +13,9 @@ class _catch extends \Spawnia\Sailor\ObjectLike
     /**
      * @param \Spawnia\Sailor\PhpKeywords\Operations\_catch\_Print\_Switch|null $print
      */
-    public static function make($print = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): self
+    public static function make(
+        $print = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
     {
         $instance = new self;