+++ examples/php-keywords/expected/Types/_new.php
@@ -12,7 +12,9 @@ class _new extends \Spawnia\Sailor\ObjectLike
     /**
      * @param float|int|null $unset
      */
-    public static function make($unset = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): self
+    public static function make(
+        $unset = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
     {
         $instance = new self;