+++ examples/polymorphic/expected/Operations/NodeWithFragments/Node/Node/Node/User.php
@@ -17,8 +17,9 @@ class User extends \Spawnia\Sailor\ObjectLike
      */
     public static function make(
         $id,
-        $name = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'
-    ): self {
+        $name = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
+    {
         $instance = new self;
 
         if ($id !== self::UNDEFINED) {