+++ examples/polymorphic/expected/Operations/PolymorphicCommonSubChildren/Sub/Nodes/Task.php
@@ -20,8 +20,9 @@ class Task extends \Spawnia\Sailor\ObjectLike
     public static function make(
         $id,
         $done,
-        $node = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'
-    ): self {
+        $node = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
+    {
         $instance = new self;
 
         if ($id !== self::UNDEFINED) {