+++ src/Operation.php
@@ -72,7 +72,7 @@ abstract class Operation implements BelongsToEndpoint
     /**
      * Send an operation through the client and return the response.
      *
-     * @param  array<int, mixed>  $args
+     * @param  array<mixed>  $args
      */
     protected static function fetchResponse(array $args): Response
     {
@@ -92,7 +92,7 @@ abstract class Operation implements BelongsToEndpoint
     }
 
     /**
-     * @param array<int, mixed> $args
+     * @param array<mixed> $args
      */
     protected static function variables(array $args): \stdClass
     {